Updated changes for Junit

+++ src/main/java/com/springboot/demo/service/impl/EmployeeServiceImpl.java
@@ -1,17 +1,15 @@
 package com.springboot.demo.service.impl;
-
 import com.springboot.demo.dao.EmployeeDao;
 import com.springboot.demo.dto.employee.EmployeeDTO;
 import com.springboot.demo.dto.employee.EmployeeListResponse;
 import com.springboot.demo.dto.employee.EmployeeResponse;
 import com.springboot.demo.entities.Employee;
 import com.springboot.demo.exception.NameAlreadyExistException;
-import com.springboot.demo.dto.BaseResponse;
+import com.springboot.demo.exception.ResourceNotFoundException;
 import com.springboot.demo.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
-
 import java.util.List;
 
 @Service
@@ -31,28 +29,29 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public EmployeeResponse addEmployee(@RequestBody EmployeeDTO employee){
+    public EmployeeResponse addEmployee(@RequestBody EmployeeDTO employee)throws NameAlreadyExistException{
         if(employeeDao.findEmployeeByName(employee.getName()) != null){
             throw new NameAlreadyExistException("Employee Name already exist");
         }else{
-//            Employee emp = new Employee(employee);
-//            Employee e = employeeDao.save(emp);
             return  new EmployeeResponse(employeeDao.save(new Employee(employee)));
         }
     }
 
     @Override
-    public Employee updateEmployee(int id, Employee employee) {
+    public EmployeeResponse updateEmployee(int id, Employee employee) {
         Employee employees = employeeDao.findById(id).get();
         employees.setName(employee.getName());
         employees.setSalary(employee.getSalary());
-        employeeDao.save(employees);
-        return employees;
+        return new EmployeeResponse(employeeDao.save(employees));
     }
 
     @Override
-    public void deleteEmployee(int id) {
+    public void deleteEmployee(int id) throws ResourceNotFoundException{
         Employee entity = employeeDao.getOne(id);
-        employeeDao.delete(entity);
+        if(employeeDao.existsById(id)){
+            employeeDao.delete(entity);
+        }else {
+            throw new ResourceNotFoundException("Employee not found for this id : " + id);
+        }
     }
 }
